# Fast ISO8601 timestamp parsing