--- a/iso8601/parse.go
+++ b/iso8601/parse.go
@@ -16,19 +16,17 @@ var (
 	errSecondOutOfRange = errors.New("second out of range")
 )
 
-// Parse parses an ISO8601 string, e.g. "2021-03-25T21:36:12Z".
+// Parse parses an ISO8601 timestamp, e.g. "2021-03-25T21:36:12Z".
 func Parse(input string) (time.Time, error) {
 	switch b := unsafeStringToBytes(input); len(b) {
-	case 20:
+	case 20: // YYYY-MM-DDTHH:MM:SSZ
 		t1 := binary.LittleEndian.Uint64(b)
 		t2 := binary.LittleEndian.Uint64(b[8:])
 		t3 := uint64(binary.LittleEndian.Uint32(b[16:]))
 
-		// Check for valid "YYYY-MM-DDTHH:MM:SSZ" separators by masking input
-		// with "    -  -  T  :  :  Z". If separators are all valid, replace
-		// them with a '0' (0x30) byte and check all bytes are now numeric.
-		// If it's not input we can handle on the fast path, pass it off to
-		// time.Parse().
+		// Check for valid separators by masking input with "    -  -  T  :  :  Z".
+		// If separators are all valid, replace them with a '0' (0x30) byte and
+		// check all bytes are now numeric.
 		if !match(t1, mask1) || !match(t2, mask2) || !match(t3, mask3) {
 			return time.Time{}, errInvalidTimestamp
 		}
@@ -55,15 +53,54 @@ func Parse(input string) (time.Time, error) {
 
 		unixSeconds := int64(daysSinceEpoch(year, month, day))*86400 + int64(hour*3600+minute*60+second)
 		return time.Unix(unixSeconds, 0), nil
-	}
 
-	return time.Parse(time.RFC3339Nano, input)
+	case 24: // YYYY-MM-DDTHH:MM:SS.MMMZ
+		t1 := binary.LittleEndian.Uint64(b)
+		t2 := binary.LittleEndian.Uint64(b[8:])
+		t4 := binary.LittleEndian.Uint64(b[16:])
+
+		// Check for valid separators by masking input with "    -  -  T  :  :  .   Z".
+		// If separators are all valid, replace them with a '0' (0x30) byte and
+		// check all bytes are now numeric.
+		if !match(t1, mask1) || !match(t2, mask2) || !match(t4, mask4) {
+			// Note that there
+			return time.Time{}, errInvalidTimestamp
+		}
+		t1 ^= replace1
+		t2 ^= replace2
+		t4 ^= replace4
+		if (nonNumeric(t1) | nonNumeric(t2) | nonNumeric(t4)) != 0 {
+			return time.Time{}, errInvalidTimestamp
+		}
+
+		t1 -= zero
+		t2 -= zero
+		t4 -= zero
+		year := (t1&0xF)*1000 + (t1>>8&0xF)*100 + (t1>>16&0xF)*10 + (t1 >> 24 & 0xF)
+		month := (t1>>40&0xF)*10 + (t1 >> 48 & 0xF)
+		day := (t2&0xF)*10 + (t2 >> 8 & 0xF)
+		hour := (t2>>24&0xF)*10 + (t2 >> 32 & 0xF)
+		minute := (t2>>48&0xF)*10 + (t2 >> 56)
+		second := (t4>>8&0xF)*10 + (t4 >> 16 & 0xF)
+		millis := (t4>>32&0xF)*100 + (t4>>40&0xF)*10 + (t4 >> 48)
+
+		if err := validate(year, month, day, hour, minute, second); err != nil {
+			return time.Time{}, err
+		}
+
+		unixSeconds := int64(daysSinceEpoch(year, month, day))*86400 + int64(hour*3600+minute*60+second)
+		return time.Unix(unixSeconds, int64(millis*1e6)), nil
+
+	default:
+		return time.Parse(time.RFC3339Nano, input)
+	}
 }
 
 const (
 	mask1 = 0x2d00002d00000000 // YYYY-MM-
 	mask2 = 0x00003a0000540000 // DDTHH:MM
 	mask3 = 0x000000005a00003a // :SSZ____
+	mask4 = 0x5a0000002e00003a // :SS.MMMZ
 
 	// Generate masks that replace the separators with a numeric byte.
 	// The input must have valid separators. XOR with the separator bytes
@@ -71,6 +108,7 @@ const (
 	replace1 = mask1 ^ 0x3000003000000000
 	replace2 = mask2 ^ 0x0000300000300000
 	replace3 = mask3 ^ 0x3030303030000030
+	replace4 = mask4 ^ 0x3000000030000030
 
 	lsb = ^uint64(0) / 255
 	msb = lsb * 0x80