--- a/iso8601/parse_test.go
+++ b/iso8601/parse_test.go
@@ -10,6 +10,7 @@ func TestParse(t *testing.T) {
 		// Fast path
 		"1987-12-16T23:45:12Z",
 		"2006-01-02T15:04:05Z",
+		"2000-02-29T23:59:59Z", // leap year
 		"2020-02-29T23:59:59Z", // leap year
 		"0000-01-01T00:00:00Z",
 		"9999-12-31T23:59:59Z",
@@ -20,13 +21,6 @@ func TestParse(t *testing.T) {
 		"2006-01-02T15:04:05.123+00:00",
 		"2006-01-02T15:04:05.123456Z",
 		"2006-01-02T15:04:05.123456789Z",
-
-		// FIXME:
-		// "2021-02-29T00:00:00Z", // not a leap year
-		// "2021-04-31T00:00:00Z", // 30 days in month
-		// "2021-06-31T00:00:00Z", // 30 days in month
-		// "2021-09-31T00:00:00Z", // 30 days in month
-		// "2021-11-31T00:00:00Z", // 30 days in month
 	} {
 		t.Run(input, func(t *testing.T) {
 			expect, err := time.Parse(time.RFC3339Nano, input)
@@ -38,6 +32,35 @@ func TestParse(t *testing.T) {
 				t.Error(err)
 			} else if !actual.Equal(expect) {
 				t.Errorf("unexpected time: %v vs expected %v", actual, expect)
+			}
+		})
+	}
+}
+
+func TestParseInvalid(t *testing.T) {
+	for _, input := range []string{
+		"1900-02-29T00:00:00Z", // not a leap year
+		"2021-02-29T00:00:00Z", // not a leap year
+		"2021-04-31T00:00:00Z", // 30 days in month
+		"2021-06-31T00:00:00Z", // 30 days in month
+		"2021-09-31T00:00:00Z", // 30 days in month
+		"2021-11-31T00:00:00Z", // 30 days in month
+		"2000-13-01T00:00:00Z", // invalid month
+		"2000-12-32T00:00:00Z", // invalid day
+		"2000-12-31T24:00:00Z", // invalid hour
+		"2000-12-31T23:60:00Z", // invalid minute
+		"2000-12-31T23:59:60Z", // invalid second
+	} {
+		t.Run(input, func(t *testing.T) {
+			ts, err := time.Parse(time.RFC3339Nano, input)
+			if err == nil {
+				t.Fatalf("expected time.Parse('%s') error, got %v", input, ts)
+			}
+			ts, actualErr := Parse(input)
+			if actualErr == nil {
+				t.Fatalf("expected Parse('%s') error %v, got %v", input, err, ts)
+			} else if err.Error() != actualErr.Error() {
+				t.Fatalf("expected Parse('%s') error %v, got %v", input, err, actualErr)
 			}
 		})
 	}