I've added a new `iso8601.Parse` function with fast paths for strings of the form `YYYY-MM-DDTHH:MM:SSZ` and`YYYY-MM-DDTHH:MM:SS.MMMZ`. If the input string isn't one of these forms, it defers to `time.Parse(time.RFC3339Nano, str)`.

It's a good deal faster than `time.Parse` with valid/invalid 20/24 byte inputs:

```
name                 old time/op    new time/op    delta
Parse-4                 218ns ± 4%      21ns ± 1%   -90.49%  (p=0.008 n=5+5)
ParseMilliseconds-4     253ns ± 0%      22ns ± 0%   -91.24%  (p=0.016 n=4+5)
ParseMicroseconds-4     256ns ± 0%     260ns ± 0%    +1.60%  (p=0.008 n=5+5)
ParseNanoseconds-4      261ns ± 0%     264ns ± 0%    +0.90%  (p=0.008 n=5+5)
ParseInvalid-4          242ns ± 0%       5ns ± 0%   -97.90%  (p=0.008 n=5+5)

name                 old alloc/op   new alloc/op   delta
Parse-4                 0.00B          0.00B           ~     (all equal)
ParseMilliseconds-4     0.00B          0.00B           ~     (all equal)
ParseMicroseconds-4     0.00B          0.00B           ~     (all equal)
ParseNanoseconds-4      0.00B          0.00B           ~     (all equal)
ParseInvalid-4          80.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
```

I've only added the two special cases given how common these forms are. It wouldn't be too difficult to add additional fast paths for the microseconds and nanosecond variants.